+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+- Added a color picker for the arrows!
+
 ## 1.2.0
 
-- Updated to work with SOTS 2.0 update
+- Now works with SOTS 2.0!
+
+## 1.3.0
+
+- Added a color picker for the Completed Arrows!